index.js in each component folder now exports default from the actual component's javascript file
