+++ src/components/Model3/Model3Section8/Model3Section8.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const index = () => {
+  return <div></div>;
+};
+
+export default index;
