+++ src/components/Model3/Model3Section8/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Model3Section8";
